Rename *.h -> *.hpp
The goal is to make GitHub show the project as consisting only of C++
files, no C files.

The .rc file is treated strangely by Vim and by Visual Studio, so I
couldn't edit it directly to account for the renamed headers. Instead, I
copied the plain text and pasted it into a new UTF-8 file.
