+++ SettingsWatchdog/stdafx.cpp
@@ -2,7 +2,7 @@
 // SettingsWatchdog.pch will be the pre-compiled header
 // stdafx.obj will contain the pre-compiled type information
 
-#include "stdafx.h"
+#include "stdafx.hpp"
 
 // TODO: reference any additional headers you need in STDAFX.H
 // and not in this file
