+++ SettingsWatchdog/stdafx.hpp
@@ -1,11 +1,11 @@
-// stdafx.h : include file for standard system include files,
+// stdafx.hpp: include file for standard system include files,
 // or project specific include files that are used frequently, but
 // are changed infrequently
 //
 
 #pragma once
 
-#include "targetver.h"
+#include "targetver.hpp"
 
 #define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
 #include <windows.h>
